requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environs==11.0.0
et_xmlfile==2.0.0
idna==3.10
marshmallow==3.23.0
numpy==2.3.0
openpyxl==3.1.5
packaging==24.1
pandas==2.3.0
paystackapi==2.1.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
tzdata==2025.2
urllib3==2.2.3
XlsxWriter==3.2.0
